Tech-sup
This project will be used by Amurt Health Care. In Amurt Health care there’re has many departments and many employees almost all of the employees are using a laptop or desktop and uses other hardware such printers and scanner. The IT Department of the Amurt (Philip Cheruyot) is the department that is in charge to maintain devices. By the use of this system, the employees' concerns or complaints will be more organized and the IT Department will monitor in the system if there's a concern from other departments who needs assistance from the IT Department (Philip Cheruyot).
Features
	Manage Clients Ticket
	Manage Categories
	Manage Users
	System Settings
	Generate CSV Reports
	Login and Logout